# Run audit on deploy