--- a/lib/assets/generateManifest.js
+++ b/lib/assets/generateManifest.js
@@ -1,11 +1,10 @@
 const fs = require('fs'),
-  glob = require('glob');
-const APP_DIR = 'app';
-const LEGACY_APP_DIR = 'marketplace_builder';
-const MODULES_DIR = 'modules';
-const appDirectory = fs.existsSync(APP_DIR) ? APP_DIR : LEGACY_APP_DIR;
+  glob = require('glob'),
+  dir = require('../directories');
 
-const serializerManifestEntry = (file) => {
+const appDirectory = fs.existsSync(dir.APP) ? dir.APP : dir.LEGACY_APP;
+
+const serializerManifestEntry = file => {
   const fileUpdatedAt = Math.floor(new Date(fs.statSync(file)['mtime']) / 1000);
   return { physical_file_path: file.replace(new RegExp(`^${appDirectory}/`), ''), updated_at: fileUpdatedAt };
 };