--- a/lib/assets/packAssets.js
+++ b/lib/assets/packAssets.js
@@ -3,16 +3,15 @@ const archiver = require('archiver-promise'),
   glob = require('glob'),
   shell = require('shelljs'),
   templates = require('../templates'),
-  settings = require('../settings');
-const APP_DIR = 'app';
-const LEGACY_APP_DIR = 'marketplace_builder';
-const MODULES_DIR = 'modules';
-const app_directory = fs.existsSync(APP_DIR) ? APP_DIR : LEGACY_APP_DIR;
+  settings = require('../settings'),
+  dir = require('../directories');
+
+const appDirectory = fs.existsSync(dir.APP) ? dir.APP : dir.LEGACY_APP;
 
 const addModulesToArchive = archive => {
-  if (!fs.existsSync(MODULES_DIR)) return true;
+  if (!fs.existsSync(dir.MODULES)) return true;
 
-  const modules = glob.sync('*/', { cwd: MODULES_DIR });
+  const modules = glob.sync('*/', { cwd: dir.MODULES });
   for (const module of modules) {
     addModuleToArchive(module, archive);
   }
@@ -23,11 +22,11 @@ const publicAssetsSameAsPrivate = (file, files) => {
 };
 
 const addModuleToArchive = (module, archive, pattern = '?(public|private)/assets/**') => {
-  const files = glob.sync(pattern, { cwd: `${MODULES_DIR}/${module}`, nodir: true });
+  const files = glob.sync(pattern, { cwd: `${dir.MODULES}/${module}`, nodir: true });
   for (const f of files) {
     if (publicAssetsSameAsPrivate(f, files)) continue;
 
-    const path = `${MODULES_DIR}/${module}/${f}`;
+    const path = `${dir.MODULES}/${module}/${f}`;
     const pathInArchive = path.replace(/(public|private)\/assets\//, '');
     archive.append(templates.fillInTemplateValues(path, settings.loadSettingsFileForModule(module)), { name: pathInArchive });
   }
@@ -41,7 +40,7 @@ const prepareDestination = (path) => {
 const packAssets = async path => {
   prepareDestination(path);
   const assetsArchive = archiver(path, { zlib: { level: 6 }, store: true });
-  assetsArchive.glob('**/**', { cwd: `${app_directory}/assets` });
+  assetsArchive.glob('**/**', { cwd: `${appDirectory}/assets` });
   addModulesToArchive(assetsArchive);
   return assetsArchive.finalize();
 };