--- a/lib/assets/packAssets.js
+++ b/lib/assets/packAssets.js
@@ -1,7 +1,7 @@
 const archiver = require('archiver-promise'),
   fs = require('fs'),
   glob = require('glob'),
-  shell = require('shelljs'),
+  shell = require('@platform-os/shelljs'),
   templates = require('../templates'),
   settings = require('../settings'),
   dir = require('../directories');