--- a/lib/audit/detailed.js
+++ b/lib/audit/detailed.js
@@ -13,7 +13,8 @@ const rules = [
   {
     glob: 'views/pages/**/*.liquid',
     test: '^enable_profiler: true$',
-    message: 'Please remove "enable_profiler: true" setting for performance reasons. \nRead more at: https://documentation.platformos.com/release-notes/15-NOVEMBER-2018#liquid-profiler'
+    message: `Please remove "enable_profiler: true" setting for performance reasons.
+      \nRead more at: https://documentation.platformos.com/release-notes/15-NOVEMBER-2018#liquid-profiler`
   },
   {
     glob: 'views/**/*.liquid',