--- a/lib/audit/filters.js
+++ b/lib/audit/filters.js
@@ -1,4 +1,4 @@
-module.exports = [
+const rules = [
   'active_class',
   'already_favorite',
   'attachments_visible_for',
@@ -59,3 +59,15 @@ module.exports = [
   'user_message_create_path',
   'widget_links'
 ];
+
+const processRules = (val, message) => {
+  return {
+    glob: '{views,authorization_policies,form_configurations,notifications}/**/*.liquid',
+    test: val,
+    message: message
+  };
+};
+
+module.exports = {
+  getRules: () => rules.map(val => processRules(`\\| ${val}`, `Deprecated "${val}" filter found.`))
+};