--- a/lib/audit/tags.js
+++ b/lib/audit/tags.js
@@ -1,4 +1,4 @@
-module.exports = [
+const rules = [
   'cache_for',
   'content_holder_tag_for_path',
   'content_holder',
@@ -25,5 +25,19 @@ module.exports = [
   'submit',
   'title',
   'transactable_type_select',
-  'will_paginate'
+  'will_paginate',
+  'query_graph',
+  'execute_query'
 ];
+
+const processRules = (val, message) => {
+  return {
+    glob: '{views,authorization_policies,form_configurations,notifications}/**/*.liquid',
+    test: val,
+    message: message
+  };
+};
+
+module.exports = {
+  getRules: () => rules.map(val => processRules(`{%-?\\s*${val}.*-?%}`, `Deprecated "${val}" tag found.`))
+};