--- a/lib/directories.js
+++ b/lib/directories.js
@@ -1,0 +1,15 @@
+const app = {
+  APP: 'app',
+  LEGACY_APP: 'marketplace_builder',
+  MODULES: 'modules',
+};
+
+const internal = {
+  TMP: '.tmp'
+};
+
+const computed = {
+  ALLOWED: [app.APP, app.LEGACY_APP, app.MODULES]
+};
+
+module.exports = Object.assign({}, app, internal, computed);