--- a/lib/prepareArchive.js
+++ b/lib/prepareArchive.js
@@ -1,5 +1,5 @@
 const fs = require('fs'),
-  shell = require('shelljs'),
+  shell = require('@platform-os/shelljs'),
   logger = require('./logger'),
   archiver = require('archiver');