--- a/lib/settings.js
+++ b/lib/settings.js
@@ -1,7 +1,6 @@
 const fs = require('fs'),
-  logger = require('./logger');
-
-const MODULES_DIR = 'modules';
+  logger = require('./logger'),
+  dir = require('./directories');
 
 const loadSettingsFile = path => {
   if (fs.existsSync(path)) {
@@ -12,7 +11,7 @@ const loadSettingsFile = path => {
 };
 
 const loadSettingsFileForModule = module => {
-  return loadSettingsFile(`${MODULES_DIR}/${module}/template-values.json`);
+  return loadSettingsFile(`${dir.MODULES}/${module}/template-values.json`);
 };
 
 const existingSettings = () => {