--- a/lib/validators/directoryEmpty.js
+++ b/lib/validators/directoryEmpty.js
@@ -1,5 +1,5 @@
 const logger = require('../logger');
-const shell = require('shelljs');
+const shell = require('@platform-os/shelljs');
 const noop = () => {};
 
 module.exports = ({ path, message = 'Directory is empty.', fail = noop }) => {