--- a/marketplace-kit-data-export.js
+++ b/marketplace-kit-data-export.js
@@ -3,7 +3,7 @@
 const program = require('commander'),
   fs = require('fs'),
   ora = require('ora'),
-  shell = require('shelljs'),
+  shell = require('@platform-os/shelljs'),
   Gateway = require('./lib/proxy'),
   logger = require('./lib/logger'),
   fetchAuthData = require('./lib/settings').fetchSettings,