--- a/marketplace-kit-deploy.js
+++ b/marketplace-kit-deploy.js
@@ -4,6 +4,7 @@ const program = require('commander'),
   fetchAuthData = require('./lib/settings').fetchSettings,
   spawn = require('child_process').spawn,
   ora = require('ora'),
+  sh = require('shelljs'),
   command = require('./lib/command'),
   logger = require('./lib/logger'),
   Gateway = require('./lib/proxy'),
@@ -20,7 +21,7 @@ const uploadArchive = (env, usingDeploymentService) => {
     archive.on('close', code => {
       if (code === 1) {
         logger.Error('Deploy failed.');
-        reject(e);
+        reject();
       }
 
       const push = spawn(command('marketplace-kit-push'), [], {
@@ -28,11 +29,11 @@ const uploadArchive = (env, usingDeploymentService) => {
         env: env
       });
 
-      push.on('close', (exitCode) => {
+      push.on('close', exitCode => {
         if (exitCode === 1) {
           logger.Error('Deploy failed.');
           reject(false);
-        } else if(exitCode === 0) {
+        } else if (exitCode === 0) {
           resolve(true);
         }
       });
@@ -40,30 +41,32 @@ const uploadArchive = (env, usingDeploymentService) => {
   });
 };
 
-const deploy = async(env, authData, params) => {
+const deploy = async (env, authData, params) => {
   await uploadArchive(env, params.directAssetsUpload);
-  if (params.directAssetsUpload){
+  if (params.directAssetsUpload) {
     const spinner = ora({ text: 'Uploading assets', stream: process.stdout, spinner: 'bouncingBar' }).start();
+
     await assets.deployAssets(new Gateway(authData));
 
     spinner.stopAndPersist().succeed('Assets uploaded');
   }
 };
 
-PARTNER_PORTAL_HOST = process.env.PARTNER_PORTAL_HOST || 'https://portal.apps.near-me.com';
-
 program
   .version(version)
   .arguments('[environment]', 'name of environment. Example: staging')
   .option('-f --force', 'force update. Removes instance-admin lock')
-  .option('-d --direct-assets-upload', 'Uploads assets straight to S3 servers')
+  .option('-d --direct-assets-upload', 'Uploads assets straight to S3 servers. It should be faster. [experimental]')
   .option('-p --partial-deploy', 'Partial deployment, does not remove data from directories missing from the build')
   .option('-c --config-file <config-file>', 'config file path', '.marketplace-kit')
   .action(async (environment, params) => {
     process.env.CONFIG_FILE_PATH = params.configFile;
+
     if (params.force) process.env.FORCE = params.force;
     if (params.partialDeploy) process.env.PARTIAL_DEPLOY = params.partialDeploy;
+
     const authData = fetchAuthData(environment, program);
+
     const env = Object.assign(process.env, {
       MARKETPLACE_EMAIL: authData.email,
       MARKETPLACE_TOKEN: authData.token,
@@ -71,12 +74,20 @@ program
       MARKETPLACE_ENV: environment
     });
 
-    try{
-      await deploy(env, authData, params);
-      process.exit(0);
-    } catch(error) {
-      process.exit(1);
-    }
+    Promise.all([
+      new Promise(resolve => {
+        // Workaround to have colors when running shell scripts via `npm`
+        logger.Info('Starting audit...');
+
+        sh.exec('FORCE_COLOR=true marketplace-kit audit', () => {
+          logger.Info('Audit done.');
+          resolve();
+        });
+      }),
+      deploy(env, authData, params)
+    ])
+      .then(() => process.exit(0))
+      .catch(() => process.exit(1));
   });
 
 program.parse(process.argv);