--- a/package.json
+++ b/package.json
@@ -40,7 +40,7 @@
     "request": "^2.88.0",
     "request-promise": "^4.2.2",
     "semver": "^5.6.0",
-    "shelljs": "^0.8.3",
+    "@platform-os/shelljs": "^0.8.3",
     "socket.io-client": "^2.1.1",
     "update-notifier": "^2.5.0",
     "uuid": "^3.3.2",