… lib/audit/rules.js
… lib/data/fetchFiles.js
… npm-shrinkwrap.json