Fixes #192