@pavelloz 2019-07-04
Performance to fix by forking shelljs and using more efficient glob engine:

Before:
```
marketplace-nearme-example|master⚡ ⇒ time marketplace-kit audit
...
marketplace-kit audit  15.35s user 5.14s system 100% cpu 20.344 total
```

After:
```
marketplace-nearme-example|master⚡ ⇒ time marketplace-kit audit
...
marketplace-kit audit  1.04s user 0.82s system 104% cpu 1.783 total
```